Changed most of the XML Input parsing. XML_Reader now -catches missing and misspelled tags -checks for bad values and  invalid data types -checks for missing tags -checks for improperly nested tags
Changes were also made to how the XML file is formatted so that it is easier for the user to enter the XML and understand what is going on. These changes are in
example2.xml
-valid set shows valid values when appropriate
-range shows valid range when appropriate
-list indicates multiple values from the valid set are appropriate
-if range or valid set is absent than any value will be accepted

Work still needs to be done. A few things are probably still missing. Need to hunt those down. Also need to work on parsing the TLE and checking its validty.